api: segments/narratives pakai float bukan Decimal di response
Decimal ter-serialize JSON sebagai string presisi-tetap ("24.00"), bikin
frontend tampil "24.00%" alih-alih "24%" -- ketahuan lewat verifikasi live
di browser (bukan cuma build hijau). Data ini presentasi, bukan angka
finansial yang butuh presisi eksak, jadi float aman dipakai -- pola yang
sama seperti Metric.value di schemas/common.py.

+++ apps/api/app/routers/narratives.py
@@ -9,7 +9,6 @@ tidak menjalankan pipeline apa pun.
 from __future__ import annotations
 
 from datetime import datetime
-from decimal import Decimal
 from uuid import UUID
 
 from fastapi import APIRouter
@@ -28,11 +27,12 @@ class NarrativeOut(BaseModel):
     code: str
     statement: str
     origin_source: SignalSource
-    volume_pct: Decimal
-    momentum_7d: Decimal
-    sentiment: Decimal | None
+    #: float, bukan Decimal — lihat catatan di routers/segments.py:SegmentOut.
+    volume_pct: float
+    momentum_7d: float
+    sentiment: float | None
     media_pickup: int
-    unclustered_pct: Decimal
+    unclustered_pct: float
     detected_at: datetime
